Create Intergalactic Elevator
The elevator has now been upgraded to operate in space! It can travel in the (x,y,z) dimensions to deliver astronauts to any destination!  (technically, it can operate through black holes, and travel through an infinite number of N dimensions!).  It now uses Unsupervised ML to sort the astronauts into clusters, before determining which of the multiple elevator shafts is quickest for them to enter.  They have also fixed the 5% reliability issue.